Change style for flex / list categories example
This is for page with leading extra span. We have developed a page and
style sheet that is 2 spans per list item, title and date, so more
semantic goodness to come. For now lets commit the pixyll perfect css
for the 3 span example. If not perfect it must be best we have and it
will be replaced eventually anyway. Refactor of css coming soon also.